Use analytics summary and graph components in revenue report
